src/backend/requirements.txt
aioredis==2.0.1
asyncpg==0.29.0
boto3==1.34.113
chromadb==0.5.5
fastapi==0.111.1
llama_index==0.10.56
openai==1.36.1
pandas==2.2.2
passlib==1.7.4
psycopg2_binary==2.9.9
pydantic==2.8.2
pymongo==4.6.3
PyPDF2==3.0.1
python-dotenv==1.0.1
python_jose==3.3.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.3
selenium==4.23.0
